backend/requirements.txt
fastapi[standard]>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.21
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
python-multipart>=0.0.18
pillow>=11.0
google-genai>=1.0
httpx>=0.28
